Adds registry component to plugin
The registry and decorator are used to keep track of which models will
be archived using django model restore. This approach was chosen so that
the entire set of models for the django site will be available when the
archive tables are constructed. This allows for useful functionalities
like applying the archive capability to any models inheriting from an
abstract model that is decorated with `soft_deletable`.